# packages/backend/src/generated/swagger.json
{
    "components": {
...
        "schemas": {
            "ApiErrorPayload": {
                "properties": {
                    "error": {
                        "properties": {
                            "data": {
                                "description": "Optional data containing details of the error"
                            },
                            "message": {
                                "type": "string",
                                "description": "A friendly message summarising the error"
                            },
                            "name": {
                                "type": "string",
                                "description": "Unique name for the type of error"
                            },
                            "statusCode": {
                                "type": "number",
                                "format": "integer",
                                "description": "HTTP status code"
                            }
                        },
                        "required": ["name", "statusCode"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["error"],
                        "nullable": false
                    }
                },
                "required": ["error", "status"],
                "type": "object",
                "description": "The Error object is returned from the api any time there is an error.\nThe message contains"
            },
...
            "SchedulerFormat": {
                "enum": ["csv", "image"],
                "type": "string"
            },
            "SchedulerCsvOptions": {
                "properties": {
                    "limit": {
                        "anyOf": [
                            {
                                "type": "number",
                                "format": "double"
                            },
                            {
                                "type": "string",
                                "enum": ["table", "all"]
                            }
                        ]
                    },
                    "formatted": {
                        "type": "boolean"
                    }
                },
                "required": ["limit", "formatted"],
                "type": "object"
            },
            "SchedulerImageOptions": {
                "properties": {},
                "type": "object"
            },
            "SchedulerOptions": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/SchedulerCsvOptions"
                    },
                    {
                        "$ref": "#/components/schemas/SchedulerImageOptions"
                    }
                ]
            },
            "SchedulerBase": {
                "properties": {
                    "options": {
                        "$ref": "#/components/schemas/SchedulerOptions"
                    },
                    "dashboardUuid": {
                        "type": "string",
                        "nullable": true
                    },
                    "savedChartUuid": {
                        "type": "string",
                        "nullable": true
                    },
                    "cron": {
                        "type": "string"
                    },
                    "format": {
                        "$ref": "#/components/schemas/SchedulerFormat"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "type": "string"
                    },
                    "schedulerUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "options",
                    "dashboardUuid",
                    "savedChartUuid",
                    "cron",
                    "format",
                    "createdBy",
                    "updatedAt",
                    "createdAt",
                    "name",
                    "schedulerUuid"
                ],
                "type": "object"
            },
            "ChartScheduler": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SchedulerBase"
                    },
                    {
                        "properties": {
                            "dashboardUuid": {
                                "type": "number",
                                "enum": [null],
                                "nullable": true
                            },
                            "savedChartUuid": {
                                "type": "string"
                            }
                        },
                        "required": ["dashboardUuid", "savedChartUuid"],
                        "type": "object"
                    }
                ]
            },
            "DashboardScheduler": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SchedulerBase"
                    },
                    {
                        "properties": {
                            "dashboardUuid": {
                                "type": "string"
                            },
                            "savedChartUuid": {
                                "type": "number",
                                "enum": [null],
                                "nullable": true
                            }
                        },
                        "required": ["dashboardUuid", "savedChartUuid"],
                        "type": "object"
                    }
                ]
            },
            "Scheduler": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/ChartScheduler"
                    },
                    {
                        "$ref": "#/components/schemas/DashboardScheduler"
                    }
                ]
            },
            "SchedulerSlackTarget": {
                "properties": {
                    "channel": {
                        "type": "string"
                    },
                    "schedulerUuid": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "schedulerSlackTargetUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "channel",
                    "schedulerUuid",
                    "updatedAt",
                    "createdAt",
                    "schedulerSlackTargetUuid"
                ],
                "type": "object"
            },
            "SchedulerEmailTarget": {
                "properties": {
                    "recipient": {
                        "type": "string"
                    },
                    "schedulerUuid": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "schedulerEmailTargetUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "recipient",
                    "schedulerUuid",
                    "updatedAt",
                    "createdAt",
                    "schedulerEmailTargetUuid"
                ],
                "type": "object"
            },
            "SchedulerAndTargets": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Scheduler"
                    },
                    {
                        "properties": {
                            "targets": {
                                "items": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/SchedulerSlackTarget"
                                        },
                                        {
                                            "$ref": "#/components/schemas/SchedulerEmailTarget"
                                        }
                                    ]
                                },
                                "type": "array"
                            }
                        },
                        "required": ["targets"],
                        "type": "object"
                    }
                ]
            },
            "SchedulerJobStatus": {
                "enum": ["scheduled", "started", "completed", "error"],
                "type": "string"
            },
            "Record_string.any_": {
                "properties": {},
                "type": "object",
                "description": "Construct a type with a set of properties K of type T"
            },
            "SchedulerLog": {
                "properties": {
                    "details": {
                        "$ref": "#/components/schemas/Record_string.any_"
                    },
                    "targetType": {
                        "type": "string",
                        "enum": ["email", "slack"]
                    },
                    "target": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/SchedulerJobStatus"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "scheduledTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "jobGroup": {
                        "type": "string"
                    },
                    "jobId": {
                        "type": "string"
                    },
                    "schedulerUuid": {
                        "type": "string"
                    },
                    "task": {
                        "type": "string",
                        "enum": [
                            "handleScheduledDelivery",
                            "sendEmailNotification",
                            "sendSlackNotification",
                            "downloadCsv",
                            "compileProject",
                            "testAndCompileProject",
                            "validateProject"
                        ]
                    }
                },
                "required": [
                    "status",
                    "createdAt",
                    "scheduledTime",
                    "jobId",
                    "task"
                ],
                "type": "object"
            },
            "SchedulerWithLogs": {
                "properties": {
                    "logs": {
                        "items": {
                            "$ref": "#/components/schemas/SchedulerLog"
                        },
                        "type": "array"
                    },
                    "dashboards": {
                        "items": {
                            "properties": {
                                "dashboardUuid": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "required": ["dashboardUuid", "name"],
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "charts": {
                        "items": {
                            "properties": {
                                "savedChartUuid": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "required": ["savedChartUuid", "name"],
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "users": {
                        "items": {
                            "properties": {
                                "userUuid": {
                                    "type": "string"
                                },
                                "lastName": {
                                    "type": "string"
                                },
                                "firstName": {
                                    "type": "string"
                                }
                            },
                            "required": ["userUuid", "lastName", "firstName"],
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "schedulers": {
                        "items": {
                            "$ref": "#/components/schemas/SchedulerAndTargets"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "logs",
                    "dashboards",
                    "charts",
                    "users",
                    "schedulers"
                ],
                "type": "object"
            },
            "ApiSchedulerLogsResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/SchedulerWithLogs"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ApiSchedulerAndTargetsResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/SchedulerAndTargets"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ScheduledJobs": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": ["id", "date"],
                "type": "object"
            },
            "ApiScheduledJobsResponse": {
                "properties": {
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/ScheduledJobs"
                        },
                        "type": "array"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ApiJobStatusResponse": {
                "properties": {
                    "results": {
                        "properties": {
                            "status": {
                                "type": "string"
                            }
                        },
                        "required": ["status"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "session_cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            },
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Value should be 'ApiKey <your key>'"
            }
        }
    },
    "info": {
        "title": "Lightdash API",
        "version": "0.714.0",
        "description": "Open API documentation for all public Lightdash API endpoints",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Lightdash Support",
            "email": "support@lightdash.com",
            "url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/schedulers/{projectUuid}/logs": {
            "get": {
                "operationId": "getSchedulerLogs",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSchedulerLogsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get scheduled logs",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/schedulers/{schedulerUuid}": {
            "get": {
                "operationId": "getScheduler",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSchedulerAndTargetsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get a scheduler",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the scheduler to update",
                        "in": "path",
                        "name": "schedulerUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "updateScheduler",
                "responses": {
                    "201": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSchedulerAndTargetsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Update a scheduler",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the scheduler to update",
                        "in": "path",
                        "name": "schedulerUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "the new scheduler data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "the new scheduler data"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteScheduler",
                "responses": {
                    "201": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {},
                                        "status": {
                                            "type": "string",
                                            "enum": ["ok"],
                                            "nullable": false
                                        }
                                    },
                                    "required": ["status"],
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Delete a scheduler",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the scheduler to delete",
                        "in": "path",
                        "name": "schedulerUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/schedulers/{schedulerUuid}/jobs": {
            "get": {
                "operationId": "getScheduledJobs",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiScheduledJobsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get scheduled jobs",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "description": "The uuid of the scheduler to update",
                        "in": "path",
                        "name": "schedulerUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/schedulers/job/{jobId}/status": {
            "get": {
                "operationId": "getSchedulerJobStatus",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiJobStatusResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get a generic job status\nThis method can be used when polling from the frontend",
                "tags": ["Schedulers"],
                "security": [],
                "parameters": [
                    {
                        "description": "the jobId for the status to check",
                        "in": "path",
                        "name": "jobId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}